+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Changed
+- the `Factor` algorithm will now factor `TrackingAssumptions` if they span
+  multiple factors of the tiling. This means that the `SplittingStrategy` is
+  removed from the tracked `StrategyPack`. It does not factor
+  `ComponentAssumptions`, so using this strategy still requires the
+  `SplittingStrategy`.
+
 ### Fixed
 - remove empty assumptions when creating extra parameters in `FusionStrategy`
 - the method `Tiling.get_genf` returns the Catalan generating function for Av(123).