+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - add the `can_be_equivalent` methods to `AddAssumptionsStrategy`,
   `SplittingStrategy`, and `FusionStrategy`.
+- added a `get_assumption` method to `Tiling`
 
 ### Fixed
 - remove empty assumptions when creating extra parameters in `FusionStrategy`