+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- add the `can_be_equivalent` methods to `AddAssumptionsStrategy`,
+  `SplittingStrategy`, and `FusionStrategy`.
+- added a `get_assumption` method to `Tiling`
+
 ### Changed
 - the `Factor` algorithm will now factor `TrackingAssumptions` if they span
   multiple factors of the tiling. This means that the `SplittingStrategy` is